For Quality Control
Run graphing.m after SungMelodytoMatrix.m without clearing the workspace. This MATLAB code graphs and saves the Treetop Envelope found in SungMelodytoMatrix.m code along with the original decibels graph as a .png file so that it can be checked that the output of the Treetop Envelope faithfully displays when the patient changes notes in his/her repetition of the melody.